Remove afterLink closures, add types, move actions, add usage explanation to the symlink trait
